Merge in COSMOSEE/base from admin_updates to master

* commit '9cba0e2392ca68298a92abfbc91f98420e7208cf':
  Remove extra alert. Fix script options
  destroy not undeploy
  Add edit Dialog
  admin updates

+++ cosmos/lib/cosmos/models/tool_model.rb
@@ -269,10 +269,12 @@ EOL
     end
 
     def undeploy
-      rubys3_client = Aws::S3::Client.new
-      prefix = "#{@folder_name}/"
-      rubys3_client.list_objects(bucket: 'tools', prefix: prefix).contents.each do |object|
-        rubys3_client.delete_object(bucket: 'tools', key: object.key)
+      if @folder_name and @folder_name.to_s.length > 0
+        rubys3_client = Aws::S3::Client.new
+        prefix = "#{@folder_name}/"
+        rubys3_client.list_objects(bucket: 'tools', prefix: prefix).contents.each do |object|
+          rubys3_client.delete_object(bucket: 'tools', key: object.key)
+        end
       end
     end
